requirements.txt
PyYAML==6.0
argparse==1.4.0